pyproject.toml
[tool.poetry]
name = "gym-saturation"
version = "0.2.6"
description = "An OpenAI Gym environment for saturation provers"
authors = ["Boris Shminke <boris@shminke.ml>"]
license = "Apache-2.0"
repository = "https://github.com/inpefess/gym-saturation"
readme = "README.rst"
classifiers=[
	"Programming Language :: Python :: 3.7",
	"Programming Language :: Python :: 3.8",
	"Programming Language :: Python :: 3.9",
	"Programming Language :: Python :: 3.10",
	"License :: OSI Approved :: Apache Software License",
	"Operating System :: OS Independent",
	"Intended Audience :: Science/Research",
	"Development Status :: 3 - Alpha",
	"Environment :: Console",
	"Natural Language :: English",
	"Topic :: Scientific/Engineering :: Artificial Intelligence",
	"Typing :: Typed"
]
include = ["gym_saturation/py.typed"]

[tool.poetry.dependencies]
python = ">= 3.7.1, < 3.11"
lark-parser = "*"
gym = "*"
importlib_resources = {version = "*", markers = "python_version < \"3.9\""}
orjson = "*"

[tool.poetry.dev-dependencies]
black = "*"
jupyterlab = "*"
mypy = "*"
pylint = "*"
pytest-coverage = "*"
pre-commit = "*"
types-dataclasses = "*"
sphinx-rtd-theme = "*"
sphinx-autodoc-typehints = "*"
flake8 = "*"

[tool.black]
line-length=79

[tool.isort]
profile = "black"
src_paths = ["gym_saturation"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
